Extract replace_sierra_ids from test_utils.
