+++ crates/sierra_generator/src/lib.rs
@@ -12,6 +12,7 @@ mod local_variables;
 mod next_statement_index_fetch;
 pub mod pre_sierra;
 mod program_generator;
+pub mod replace_ids;
 mod resolve_labels;
 mod specialization_context;
 mod store_variables;
